düzenleme

+++ Builder/Model/ProductViewModel.cs
@@ -10,8 +10,8 @@ namespace Builder.Model
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int UnitPrice { get; set; }
-        public int DiscountedPrice { get; set; }
+        public float UnitPrice { get; set; }
+        public float DiscountedPrice { get; set; }
         public string Category { get; set; }
         public int Discount { get; set; }
         public bool DiscountApplied { get; set; }
